requirements.txt
requests==2.*
python-dotenv
geopy
folium
charset-normalizer